+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rewrote `README.md`.
+- The dependency on node-rs/crc32 has been replaced by one on aws-crypto/crc32c,
+  a pure-JS implementation of CRC32C.
 
 ## [3.6.0] - 2024-01-12